Added BitSet.h, fixing errors related to bitset

+++ org/antlr/v4/runtime/UnbufferedCharStream.h
@@ -173,7 +173,7 @@ namespace org {
                     /// </summary>
                     virtual void seek(int index) override;
 
-                    virtual int size() override;
+                    virtual size_t size() override;
 
                     virtual std::string getSourceName() override;
